// client-vue/src/app/store/modules/activity/mutation-types.ts
export enum MutationTypes {
  SET_ACTIVITIES = 'SET_ACTIVITIES',
  SET_ACTIVITY = 'SET_ACTIVITY',
  SET_SELECTED_ACTIVITY = 'SET_SELECTED_ACTIVITY',
  SET_EDIT_MODE = 'SET_EDIT_MODE',
  SET_IS_LOADING = 'SET_IS_LOADING',
  SET_IS_LOADING_INITIAL = 'SET_IS_LOADING_INITIAL',
  REMOVE_ACTIVITY = 'REMOVE_ACTIVITY',
}